Refacto and implement tests
