+++ cmd/collector/app/zipkin/http_handler.go
@@ -24,7 +24,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/apache/thrift/lib/go/thrift"
 	"github.com/go-openapi/loads"
 	"github.com/go-openapi/strfmt"
 	"github.com/go-openapi/swag"
@@ -32,6 +31,7 @@ import (
 	tchanThrift "github.com/uber/tchannel-go/thrift"
 
 	"github.com/jaegertracing/jaeger/cmd/collector/app"
+	"github.com/jaegertracing/jaeger/model/converter/thrift/zipkin"
 	"github.com/jaegertracing/jaeger/swagger-gen/models"
 	"github.com/jaegertracing/jaeger/swagger-gen/restapi"
 	"github.com/jaegertracing/jaeger/swagger-gen/restapi/operations"
@@ -89,7 +89,7 @@ func (aH *APIHandler) saveSpans(w http.ResponseWriter, r *http.Request) {
 
 	var tSpans []*zipkincore.Span
 	if contentType == "application/x-thrift" {
-		tSpans, err = deserializeThrift(bodyBytes)
+		tSpans, err = zipkin.DeserializeThrift(bodyBytes)
 	} else if contentType == "application/json" {
 		tSpans, err = DeserializeJSON(bodyBytes)
 	} else {
@@ -180,28 +180,4 @@ func (aH *APIHandler) saveThriftSpans(tSpans []*zipkincore.Span) error {
 		}
 	}
 	return nil
-}
-
-func deserializeThrift(b []byte) ([]*zipkincore.Span, error) {
-	buffer := thrift.NewTMemoryBuffer()
-	buffer.Write(b)
-
-	transport := thrift.NewTBinaryProtocolTransport(buffer)
-	_, size, err := transport.ReadListBegin() // Ignore the returned element type
-	if err != nil {
-		return nil, err
-	}
-
-	// We don't depend on the size returned by ReadListBegin to preallocate the array because it
-	// sometimes returns a nil error on bad input and provides an unreasonably large int for size
-	var spans []*zipkincore.Span
-	for i := 0; i < size; i++ {
-		zs := &zipkincore.Span{}
-		if err = zs.Read(transport); err != nil {
-			return nil, err
-		}
-		spans = append(spans, zs)
-	}
-
-	return spans, nil
 }
